ml_models_inference/ner_models/address_recognition_flair/requirements.txt
flair
transformers
scikit-learn
pandas
numpy
boto3
flask
python-dotenv